requirement.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==4.1.1
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.41.5
PyJWT==2.9.0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
